refactor(frontend): extract lecture watch playback hook
